Showtime /simplify: extract slotsFromOptions, parallelize market fetch
- slotsFromOptions shared helper (ShowtimeBubbles) replaces the duplicated
  options+metadata -> ShowtimeSlot[] builder in ShowtimeBallotSection + ShowtimeResults
- GET /nearby fetches covered markets concurrently (asyncio.gather) instead of
  sequentially
- drop unused HTTPException import
